Minor changes to test 7

+++ QuantLib/Examples/OpenCL/mcsimulation.cpp
@@ -79,48 +79,48 @@ inline void MersenneTwister_Bulk(float* randOutput,
 // An implementation of Peter J. Acklam's inverse cumulative normal distribution function
 // Generates one random variable
 inline float AcklamInvCND(float input) {
-    const float a1 = -39.6968302866538;
-    const float a2 = 220.946098424521;
-    const float a3 = -275.928510446969;
-    const float a4 = 138.357751867269;
-    const float a5 = -30.6647980661472;
-    const float a6 = 2.50662827745924;
-    const float b1 = -54.4760987982241;
-    const float b2 = 161.585836858041;
-    const float b3 = -155.698979859887;
-    const float b4 = 66.8013118877197;
-    const float b5 = -13.2806815528857;
-    const float c1 = -7.78489400243029E-03;
-    const float c2 = -0.322396458041136;
-    const float c3 = -2.40075827716184;
-    const float c4 = -2.54973253934373;
-    const float c5 = 4.37466414146497;
-    const float c6 = 2.93816398269878;
-    const float d1 = 7.78469570904146E-03;
-    const float d2 = 0.32246712907004;
-    const float d3 = 2.445134137143;
-    const float d4 = 3.75440866190742;
-    const float p_low = 0.02425;
-    const float p_high = 1.0 - p_low;
+    const float a1 = -39.6968302866538f;
+    const float a2 = 220.946098424521f;
+    const float a3 = -275.928510446969f;
+    const float a4 = 138.357751867269f;
+    const float a5 = -30.6647980661472f;
+    const float a6 = 2.50662827745924f;
+    const float b1 = -54.4760987982241f;
+    const float b2 = 161.585836858041f;
+    const float b3 = -155.698979859887f;
+    const float b4 = 66.8013118877197f;
+    const float b5 = -13.2806815528857f;
+    const float c1 = -7.78489400243029E-03f;
+    const float c2 = -0.322396458041136f;
+    const float c3 = -2.40075827716184f;
+    const float c4 = -2.54973253934373f;
+    const float c5 = 4.37466414146497f;
+    const float c6 = 2.93816398269878f;
+    const float d1 = 7.78469570904146E-03f;
+    const float d2 = 0.32246712907004f;
+    const float d3 = 2.445134137143f;
+    const float d4 = 3.75440866190742f;
+    const float p_low = 0.02425f;
+    const float p_high = 1.0f - p_low;
     float z, R;
 
 	if((input) <= 0 || (input) >= 1.0)
 		(input) = (float)0x7FFFFFFF;
 
 	if((input) < p_low){
-		z = sqrt(-2.0 * log((input)));
+		z = (float)sqrt(-2.0 * log((input)));
 		z = (((((c1 * z + c2) * z + c3) * z + c4) * z + c5) * z + c6) /
-			((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0);
+			((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0f);
 	}else{
 		if((input) > p_high){
-			z = sqrt(-2.0 * log(1.0 - (input)));
+			z = (float)sqrt(-2.0 * log(1.0 - (input)));
 			z = -(((((c1 * z + c2) * z + c3) * z + c4) * z + c5) * z + c6) /
-				 ((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0);
+				 ((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0f);
 		} else {
-			z = (input) - 0.5;
+			z = (input) - 0.5f;
 			R = z * z;
 			z = (((((a1 * R + a2) * R + a3) * R + a4) * R + a5) * R + a6) * z /
-				(((((b1 * R + b2) * R + b3) * R + b4) * R + b5) * R + 1.0);
+				(((((b1 * R + b2) * R + b3) * R + b4) * R + b5) * R + 1.0f);
 		}
 	}
 
@@ -130,29 +130,29 @@ inline float AcklamInvCND(float input) {
 // An implementation of Peter J. Acklam's inverse cumulative normal distribution function
 // Generates bufferLength random variables
 inline void AcklamInvCND_Bulk(float* randOutput, const size_t bufferLength) {
-    const float a1 = -39.6968302866538;
-    const float a2 = 220.946098424521;
-    const float a3 = -275.928510446969;
-    const float a4 = 138.357751867269;
-    const float a5 = -30.6647980661472;
-    const float a6 = 2.50662827745924;
-    const float b1 = -54.4760987982241;
-    const float b2 = 161.585836858041;
-    const float b3 = -155.698979859887;
-    const float b4 = 66.8013118877197;
-    const float b5 = -13.2806815528857;
-    const float c1 = -7.78489400243029E-03;
-    const float c2 = -0.322396458041136;
-    const float c3 = -2.40075827716184;
-    const float c4 = -2.54973253934373;
-    const float c5 = 4.37466414146497;
-    const float c6 = 2.93816398269878;
-    const float d1 = 7.78469570904146E-03;
-    const float d2 = 0.32246712907004;
-    const float d3 = 2.445134137143;
-    const float d4 = 3.75440866190742;
-    const float p_low = 0.02425;
-    const float p_high = 1.0 - p_low;
+    const float a1 = -39.6968302866538f;
+    const float a2 = 220.946098424521f;
+    const float a3 = -275.928510446969f;
+    const float a4 = 138.357751867269f;
+    const float a5 = -30.6647980661472f;
+    const float a6 = 2.50662827745924f;
+    const float b1 = -54.4760987982241f;
+    const float b2 = 161.585836858041f;
+    const float b3 = -155.698979859887f;
+    const float b4 = 66.8013118877197f;
+    const float b5 = -13.2806815528857f;
+    const float c1 = -7.78489400243029E-03f;
+    const float c2 = -0.322396458041136f;
+    const float c3 = -2.40075827716184f;
+    const float c4 = -2.54973253934373f;
+    const float c5 = 4.37466414146497f;
+    const float c6 = 2.93816398269878f;
+    const float d1 = 7.78469570904146E-03f;
+    const float d2 = 0.32246712907004f;
+    const float d3 = 2.445134137143f;
+    const float d4 = 3.75440866190742f;
+    const float p_low = 0.02425f;
+    const float p_high = 1.0f - p_low;
     float z, R;
 	float* target;
 
@@ -163,19 +163,19 @@ inline void AcklamInvCND_Bulk(float* randOutput, const size_t bufferLength) {
 			(*target) = (float)0x7FFFFFFF;
 
 		if((*target) < p_low){
-			z = sqrt(-2.0 * log((*target)));
+			z = (float)sqrt(-2.0 * log((*target)));
 			z = (((((c1 * z + c2) * z + c3) * z + c4) * z + c5) * z + c6) /
-				((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0);
+				((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0f);
 		}else{
 			if((*target) > p_high){
-				z = sqrt(-2.0 * log(1.0 - (*target)));
+				z = (float)sqrt(-2.0 * log(1.0 - (*target)));
 				z = -(((((c1 * z + c2) * z + c3) * z + c4) * z + c5) * z + c6) /
-					 ((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0);
+					 ((((d1 * z + d2) * z + d3) * z + d4) * z + 1.0f);
 			} else {
-				z = (*target) - 0.5;
+				z = (*target) - 0.5f;
 				R = z * z;
 				z = (((((a1 * R + a2) * R + a3) * R + a4) * R + a5) * R + a6) * z /
-					(((((b1 * R + b2) * R + b3) * R + b4) * R + b5) * R + 1.0);
+					(((((b1 * R + b2) * R + b3) * R + b4) * R + b5) * R + 1.0f);
 			}
 		}
